requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF Processing
pypdf==4.0.1

# LLM Integration
openai==1.3.9
requests==2.31.0

# Database
sqlalchemy==2.0.35
sqlalchemy-utils==0.41.1

# Utilities
python-dotenv==1.0.0
pydantic==2.8.0
pydantic-settings==2.1.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1 